requirements.txt
alabaster==0.7.12
alembic==1.5.8
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.3.4
astroid==2.5
async-generator==1.10
attrs==20.3.0
authentication==1.1.0
autopep8==1.5.6
Babel==2.8.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.2
djangorestframework==3.12.4
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
environs==9.3.2
evdev==1.3.0
Faker==6.1.1
fasteners==0.14.1
flake8==3.9.0
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.0.0
httplib2==0.14.0
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
ipdb==0.13.7
ipykernel==5.3.4
ipyparallel==6.3.0
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.13
isort==5.7.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.5.2
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
matplotlib==3.4.2
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
multibootusb==9.2.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
netifaces==0.10.4
nose==1.3.7
notebook==6.1.4
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
packaging==20.9
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.7.1
passlib==1.7.4
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.6.1
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.10.0
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pycups==1.9.73
pyflakes==2.3.1
pygame==2.0.0
Pygments==2.7.2
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.17.3
pytest==6.2.3
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.1
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
redis==3.5.3
reportlab==3.5.34
requests==2.24.0
requests-unixsocket==0.2.0
response==0.5.0
rope==0.18.0
scipy==1.6.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
serial==0.0.97
simplejson==3.16.0
sip==4.19.21
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.7
sqlparse==0.4.1
systemd-python==234
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
tornado==6.0.4
traitlets==5.0.5
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.11
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
xkit==0.0.0
yapf==0.30.0
